Address signedness warnings (#34)

+++ src/geometry.cpp
@@ -46,7 +46,7 @@ const std::array<v512, 64> SUPERPIECE_INVERSE_RAYS_AVX2_TABLE = []() {
     for (u8 sq = 0; sq < 64; sq++) {
         u8                 esq = internal::expand_sq(Square{sq});
         std::array<u8, 64> b;
-        for (int i = 0; i < 64; i++) {
+        for (usize i = 0; i < 64; i++) {
             u8 value = BASE[AVX2_OFFSETS[i] - esq];
             b[i]     = value;
         }
@@ -92,7 +92,7 @@ const std::array<v512, 64> PIECE_MOVES_AVX2_TABLE = []() {
     for (u8 sq = 0; sq < 64; sq++) {
         u8                 esq = internal::expand_sq(Square{sq});
         std::array<u8, 64> b;
-        for (int i = 0; i < 64; i++) {
+        for (usize i = 0; i < 64; i++) {
             b[i] = BASE[AVX2_OFFSETS[i] - esq];
         }
         table[sq] = v512{b};
